Move core auth mixins to their own file
